% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.metadata.R
\name{extract.metadata}
\alias{extract.metadata}
\title{read metadata from .sig files}
\usage{
extract.metadata(file, instrument = "SVC")
}
\arguments{
\item{file}{A character string specifying the path to the file.}

\item{instrument}{A character string specifying the type of instrument. Default is "SVC".}
}
\value{
A data frame with two columns:
\describe{
\item{Key}{Character vector representing the metadata keys.}
\item{Value}{Character vector representing the metadata values.}
}
}
\description{
Extract metadata from .sig file
This function reads a file generated by an instrument (default is SVC), extracts metadata, and returns it as a data frame.
}
\details{
The function reads the content of the specified file line by line. It splits each line into key-value pairs based on the "=" delimiter and excludes lines with the key "data". The extracted key-value pairs are stored in a data frame.
}
\examples{
\dontrun{
metadata <- extract.metadata("/path/to/your/file.sig", instrument = "SVC")
print(metadata)
}

}
\author{
Methun George, Steffen Neumann
}
