% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.sig.R
\name{read.sig}
\alias{read.sig}
\title{Read .sig Files from SVC}
\usage{
read.sig(file_path)
}
\arguments{
\item{file_path}{A character string specifying the path to the .sig file.}
}
\value{
A data frame with the following columns:
\describe{
  \item{Wavelengths (nm)}{Numeric vector representing the wavelengths in nanometers.}
  \item{Reference Values}{Numeric vector of reference values.}
  \item{Target Values}{Numeric vector of target values.}
  \item{Reflectance (\%)}{Numeric vector of reflectance percentages.}
}
}
\description{
Function to read .sig files and extract spectral data.
}
\details{
This function reads a .sig file generated by a NIRS machine, extracts the spectral data, and returns it as a data frame.

The function first checks if the file exists. If it doesn't exist, an error is thrown. The file is then read line by line, and the spectral data is extracted from the line following "data=". The extracted data is then converted into a data frame with appropriate column names.
}
\examples{
\dontrun{
file_path <- "/path/to/your/file.sig"
data <- read.sig(file_path)
print(head(data))
}
}
\author{
Methun George, Steffen Neumann
}
