% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_summarizedexperiment_asd.R
\name{read_summarizedexperiment_asd}
\alias{read_summarizedexperiment_asd}
\title{Read the ASD files and return as SummarizedExperiment object.}
\arguments{
\item{path}{A character string specifying the path to a directory containing the ".asd" files or a single
".asd" file.}
}
\value{
A `SummarizedExperiment` object containing:
  \itemize{
    \item{\code{assays}}{A list with a single assay matrix named \code{counts},
      where rows represent wavelengths and columns represent different samples.}
    \item{\code{colData}}{A `DataFrame` with sample metadata, including the sample names.}
    \item{\code{rowData}}{A `DataFrame` with wavelength information.}
  }
}
\description{
Read ASD Spectral Files and Create a SummarizedExperiment Object
}
\details{
This function will read a single or multiple ASD files and extracts the spectral data
and associated metadata, combines them into a "SummarizedExperiment" object. This object
contains the spectral data as an assay matrix and metadata about the samples and wavelengths.


This function uses 2 functions such as "read.asd" and "extract.metadata" from the package "FieldSpectra"
to read the data and metadata from the given files. This function will then uses the data from the files
and make a SummarizedExperiment object out of it which can be used for further data analysis.
}
\examples{
# path <- "path/to/your/asd/files"
# se <- read_asd_summar(path)
# se
# colnames(se)
# rownames(se)

}
\author{
Methun George, Dr.Steffen Neumann
}
