% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_summarizedexperiment_sig.R
\name{read_summarizedexperiment_sig}
\alias{read_summarizedexperiment_sig}
\title{Read the SIG files and return as SummarizedExperiment object.}
\arguments{
\item{path}{A character vector of file paths or a directory path. If a directory path is provided, all .sig files
within the directory are read. If a file path is provided, only that file is read.}
}
\value{
A \code{SummarizedExperiment} object where:
\item{assays}{A matrix of reflectance data, with wavelengths as row names and sample names as column names.}
\item{colData}{A \code{DataFrame} with sample names.}
\item{rowData}{A \code{DataFrame} with wavelengths.}
}
\description{
Read .sig Files and Create SummarizedExperiment Object
}
\details{
This function reads .sig files from a specified directory or a file path, extracts wavelength and reflectance data,
and combines them into a \code{SummarizedExperiment} object. Each .sig file is expected to contain reflectance data
and a sample name in its content.


Each .sig file is read line by line. The function expects each file to contain:
\itemize{
  \item A line starting with "name=" indicating the sample name.
  \item A section starting with "data=" followed by tabular data where the first column is wavelengths and the fourth
  column is reflectance.
}
The combined data from all files is merged by wavelengths. Missing data is handled by filling with NA.
}
\examples{
# Example usage:
path <- "/path/to/sig/files"
se <- read_summarizedexperiment_asd(path)
print(se)

}
\author{
Methun George, Dr.Steffen Neumann
}
